utilities/autonomous-dream-injection/requirements.txt
llama-cpp-python>=0.3.0
huggingface-hub>=0.20.0
typing-extensions>=4.5.0
numpy>=1.20.0
